Integrate Prisma for database management, bcrypt for password hashing and dotenv for environment variables
Prisma was added as a client to facilitate database management operations. Additional dependencies "@prisma/client" and "prisma" were added in the package.json. New files for managing Prisma service in the database were created.

A new user repository was created, including the repository interface. User entity was changed to add the ability to hash a password with bcrypt property and compare with plain text.

An important change was the use of dotenv to manage environment variables such as salt for bcrypt. Dotenv was also added in the dependencies. Corresponding changes for environment variable usage were committed to users.service.ts, users.controller.ts, and user.entity.ts.

Environment variable for SALT was added in a new '.env' file. Remember to add '.env' in the '.gitignore' to avoid committing sensitive information.

Also, linting tools were added for better code quality maintenance, and necessary linting changes have been made to the code.

As part of this commit, a new "generate" script was also added to generate Prisma client.

Database configuration and table creation were handled using Prisma migration files added in 'prisma' directory.

Please note that these changes require Node version >=16.13 as specified in the engines properties of prisma and @prisma/client.